+++ CHANGELOG.md
@@ -23,9 +23,21 @@
   l22 が返すのは login, password, uhour のみ、そこを変えないと改良ならない。
   あるいは /admin/:user を作るか。
 - VScode で cljs repl につながらなくなった。以前はどうだったか？
-- plot の左マージン
 - hiccup.page/html5 は警告されない。hiccup.form/form-to は警告される理由。
+- /todays に today's total を表示。
 
+## 1.18.4 - 2023-05-16
+- FIXME: plot 中で (map :pt data) は役割分担としておかしい。
+
+## 1.18.3 - 2023-05-16
+- todays, total 入れ替え
+
+## 1.18.2 - 2023-05-16
+ダミーリリース。「ブラウザの履歴をクリア」のチェック。
+
+## 1.18.1 - 2023-05-16
+- plot todays when (< 9 (count todays))
+- /todays に (headline 7)
 
 ## 1.18.0 - 2023-05-16
 - double plots. need adjust margin and position